bugfix (fixes #1635): use journald_test_cursor() from systemd-journal API for additional cursor test from syslog-ng.presit file
short bug description: when syslog-ng checks cursor position from
syslog-ng.persist file it does not check if cursor is "in the future" with
respect to the system.  This causes logging glitch when syslog-ng is started
(syslog-ng waits for the "time to come" to continue logging) when there is no
systemd journal entries with the "future time". This is because
journald_seek_cursors() seeks to the next closest entry (in terms of time) if
no entry matches.